refactor(api): move HookHandler trait and HookResult to zeroclaw-api
Moves `HookHandler` and `HookResult` from `zeroclaw-runtime` to
`zeroclaw-api` to break the dependency inversion where core API types
(depended on by all crates) could not reference hook abstractions
(defined in a downstream implementation crate).

This keeps the trait definitions in the lightweight API layer where
other fundamental abstractions (ModelProvider, Channel, Tool, etc.)
already live, and opens the door for exposing `HookHandler` as a
`PluginCapability` via `zeroclaw-plugins` (WASM) in the future.

`zeroclaw-runtime` re-exports these types for backward compatibility.
